travel-assistant-bot/requirements.txt
python-telegram-bot==21.10
groq==0.13.0
python-dotenv==1.0.1
